Additional information regarding this project
Dataset_14-dat_AA. is the original file with a lot of missing values and a lot of columns. depression symptoms.ipynb is the Python file where data cleaning and data modeling were carried out, followed by Principal Component Analysis (PCA). Following the data modeling, the number of columns was reduced to less than 10, with a few missing values in one of the columns, which were further taken care of in PowerBI. PowerBI used the semi-cleaned dataset depression_symptoms.csv obtained from Python for further analysis.